+++ src/config.rs
@@ -3,9 +3,10 @@ use std::{
 	sync::OnceLock,
 };
 
-use entity::Config;
+pub use entity::Config;
 use realme::{
 	Adaptor,
+	CmdSource,
 	FileSource,
 	Realme,
 	TomlParser,
@@ -20,9 +21,12 @@ mod entity;
 
 pub static CFG: OnceLock<Config> = OnceLock::new();
 
-pub fn initialize_config(path: &'static Path) -> Result<&'static Config> {
+pub fn initialize_config(
+	path: &Path,
+	options: &str,
+) -> Result<&'static Config> {
 	let config = CFG.get_or_init(|| {
-		Config::load_config(path).unwrap_or_else(|e| {
+		Config::load_config(path, options).unwrap_or_else(|e| {
 			eprintln!("Load config err: {e}");
 			tracing::error!("Load config err: {e}");
 			std::process::exit(78);
@@ -36,13 +40,15 @@ pub fn get_config() -> &'static Config {
 }
 
 impl Config {
-	pub fn load_config(path: &'static Path) -> Result<Self> {
+	pub fn load_config(path: &Path, options: &str) -> Result<Self> {
 		let realme = Realme::builder()
-			.load(Adaptor::new(FileSource::<TomlParser, _>::new(path)))
+			.load(Adaptor::new(FileSource::<TomlParser>::new(path)).priority(2))
+			.load(
+				Adaptor::new(CmdSource::<TomlParser>::new(options)).priority(1),
+			)
 			.build()
 			.map_err(|e| Error::Config(e.to_string()))?;
 
-		eprintln!("realme: {realme:#?}");
 		let config = realme
 			.try_deserialize()
 			.map_err(|e| Error::Config(e.to_string()))?;
@@ -52,11 +58,13 @@ impl Config {
 
 #[cfg(test)]
 mod tests {
+	use std::path::Path;
+
 	use super::*;
 
 	#[test]
 	fn test_config() {
-		let res = initialize_config(Path::new("./config/rinkle.toml"));
+		let res = initialize_config(Path::new("./config/rinkle.toml"), "");
 		assert!(res.is_ok());
 	}
 }