+++ src/lib.rs
@@ -2,3 +2,5 @@ pub mod cli;
 pub mod config;
 pub mod error;
 pub mod prelude;
+pub mod rinkle;
+pub mod utils;