+++ src/prelude.rs
@@ -1,5 +1,12 @@
 //! Prelude for the crate
-pub use crate::error::*;
+pub use crate::{
+	config::{
+		Config,
+		get_config,
+	},
+	error::*,
+	rinkle::Rinkle,
+};
 
 #[allow(dead_code)]
 /// Generic Wrapper tuple struct for newtype patterns